+++ bem/src/assembly/batched.rs
@@ -30,6 +30,13 @@ use rlst_dense::{
 use rlst_sparse::sparse::csr_mat::CsrMatrix;
 use std::collections::HashMap;
 
+mod adjoint_double_layer;
+mod double_layer;
+mod single_layer;
+pub use single_layer::{HelmholtzSingleLayerAssembler, LaplaceSingleLayerAssembler};
+pub use double_layer::{HelmholtzDoubleLayerAssembler, LaplaceDoubleLayerAssembler};
+pub use adjoint_double_layer::{HelmholtzAdjointDoubleLayerAssembler, LaplaceAdjointDoubleLayerAssembler};
+
 type RlstArray<T, const DIM: usize> = Array<T, BaseArray<T, VectorContainer<T>, DIM>, DIM>;
 
 fn equal_grids<TestGrid: GridType, TrialGrid: GridType>(
@@ -1178,313 +1185,6 @@ pub trait BatchedAssembler: Sync {
                 assert_eq!(r, numtasks);
             }
         }
-    }
-}
-
-/// Assembler for a Laplace single layer operator
-pub struct LaplaceSingleLayerAssembler<T: RlstScalar> {
-    kernel: Laplace3dKernel<T>,
-}
-impl<T: RlstScalar> Default for LaplaceSingleLayerAssembler<T> {
-    fn default() -> Self {
-        Self {
-            kernel: Laplace3dKernel::<T>::new(),
-        }
-    }
-}
-impl<T: RlstScalar> BatchedAssembler for LaplaceSingleLayerAssembler<T> {
-    const DERIV_SIZE: usize = 1;
-    type RealT = T::Real;
-    type T = T;
-    unsafe fn singular_kernel_value(
-        &self,
-        k: &RlstArray<Self::T, 2>,
-        _test_normals: &RlstArray<Self::RealT, 2>,
-        _trial_normals: &RlstArray<Self::RealT, 2>,
-        index: usize,
-    ) -> Self::T {
-        *k.get_unchecked([0, index])
-    }
-    unsafe fn nonsingular_kernel_value(
-        &self,
-        k: &RlstArray<Self::T, 3>,
-        _test_normals: &RlstArray<Self::RealT, 2>,
-        _trial_normals: &RlstArray<Self::RealT, 2>,
-        test_index: usize,
-        trial_index: usize,
-    ) -> Self::T {
-        *k.get_unchecked([test_index, 0, trial_index])
-    }
-    fn kernel_assemble_diagonal_st(
-        &self,
-        sources: &[Self::RealT],
-        targets: &[Self::RealT],
-        result: &mut [Self::T],
-    ) {
-        self.kernel
-            .assemble_diagonal_st(EvalType::Value, sources, targets, result);
-    }
-    fn kernel_assemble_st(
-        &self,
-        sources: &[Self::RealT],
-        targets: &[Self::RealT],
-        result: &mut [Self::T],
-    ) {
-        self.kernel
-            .assemble_st(EvalType::Value, sources, targets, result);
-    }
-}
-
-/// Assembler for a Helmholtz single layer boundary operator
-pub struct HelmholtzSingleLayerAssembler<T: RlstScalar<Complex = T>> {
-    kernel: Helmholtz3dKernel<T>,
-}
-impl<T: RlstScalar<Complex = T>> HelmholtzSingleLayerAssembler<T> {
-    /// Create a new assembler
-    pub fn new(wavenumber: T::Real) -> Self {
-        Self {
-            kernel: Helmholtz3dKernel::<T>::new(wavenumber),
-        }
-    }
-}
-impl<T: RlstScalar<Complex = T>> BatchedAssembler for HelmholtzSingleLayerAssembler<T> {
-    const DERIV_SIZE: usize = 1;
-    type RealT = T::Real;
-    type T = T;
-    unsafe fn singular_kernel_value(
-        &self,
-        k: &RlstArray<Self::T, 2>,
-        _test_normals: &RlstArray<Self::RealT, 2>,
-        _trial_normals: &RlstArray<Self::RealT, 2>,
-        index: usize,
-    ) -> Self::T {
-        *k.get_unchecked([0, index])
-    }
-    unsafe fn nonsingular_kernel_value(
-        &self,
-        k: &RlstArray<Self::T, 3>,
-        _test_normals: &RlstArray<Self::RealT, 2>,
-        _trial_normals: &RlstArray<Self::RealT, 2>,
-        test_index: usize,
-        trial_index: usize,
-    ) -> Self::T {
-        *k.get_unchecked([test_index, 0, trial_index])
-    }
-    fn kernel_assemble_diagonal_st(
-        &self,
-        sources: &[Self::RealT],
-        targets: &[Self::RealT],
-        result: &mut [Self::T],
-    ) {
-        self.kernel
-            .assemble_diagonal_st(EvalType::Value, sources, targets, result);
-    }
-    fn kernel_assemble_st(
-        &self,
-        sources: &[Self::RealT],
-        targets: &[Self::RealT],
-        result: &mut [Self::T],
-    ) {
-        self.kernel
-            .assemble_st(EvalType::Value, sources, targets, result);
-    }
-}
-
-/// Assembler for a Laplace double layer operator
-pub struct LaplaceDoubleLayerAssembler<T: RlstScalar> {
-    kernel: Laplace3dKernel<T>,
-}
-impl<T: RlstScalar> Default for LaplaceDoubleLayerAssembler<T> {
-    fn default() -> Self {
-        Self {
-            kernel: Laplace3dKernel::<T>::new(),
-        }
-    }
-}
-impl<T: RlstScalar> BatchedAssembler for LaplaceDoubleLayerAssembler<T> {
-    const DERIV_SIZE: usize = 4;
-    type RealT = T::Real;
-    type T = T;
-    unsafe fn singular_kernel_value(
-        &self,
-        k: &RlstArray<Self::T, 2>,
-        _test_normals: &RlstArray<Self::RealT, 2>,
-        trial_normals: &RlstArray<Self::RealT, 2>,
-        index: usize,
-    ) -> Self::T {
-        *k.get_unchecked([1, index])
-            * num::cast::<Self::RealT, Self::T>(*trial_normals.get_unchecked([index, 0])).unwrap()
-            + *k.get_unchecked([2, index])
-                * num::cast::<Self::RealT, Self::T>(*trial_normals.get_unchecked([index, 1]))
-                    .unwrap()
-            + *k.get_unchecked([3, index])
-                * num::cast::<Self::RealT, Self::T>(*trial_normals.get_unchecked([index, 2]))
-                    .unwrap()
-    }
-    unsafe fn nonsingular_kernel_value(
-        &self,
-        k: &RlstArray<Self::T, 3>,
-        _test_normals: &RlstArray<Self::RealT, 2>,
-        trial_normals: &RlstArray<Self::RealT, 2>,
-        test_index: usize,
-        trial_index: usize,
-    ) -> Self::T {
-        *k.get_unchecked([test_index, 1, trial_index])
-            * num::cast::<Self::RealT, Self::T>(*trial_normals.get_unchecked([trial_index, 0]))
-                .unwrap()
-            + *k.get_unchecked([test_index, 2, trial_index])
-                * num::cast::<Self::RealT, Self::T>(*trial_normals.get_unchecked([trial_index, 1]))
-                    .unwrap()
-            + *k.get_unchecked([test_index, 3, trial_index])
-                * num::cast::<Self::RealT, Self::T>(*trial_normals.get_unchecked([trial_index, 2]))
-                    .unwrap()
-    }
-    fn kernel_assemble_diagonal_st(
-        &self,
-        sources: &[Self::RealT],
-        targets: &[Self::RealT],
-        result: &mut [Self::T],
-    ) {
-        self.kernel
-            .assemble_diagonal_st(EvalType::ValueDeriv, sources, targets, result);
-    }
-    fn kernel_assemble_st(
-        &self,
-        sources: &[Self::RealT],
-        targets: &[Self::RealT],
-        result: &mut [Self::T],
-    ) {
-        self.kernel
-            .assemble_st(EvalType::ValueDeriv, sources, targets, result);
-    }
-}
-
-/// Assembler for a Helmholtz double layer boundary operator
-pub struct HelmholtzDoubleLayerAssembler<T: RlstScalar<Complex = T>> {
-    kernel: Helmholtz3dKernel<T>,
-}
-impl<T: RlstScalar<Complex = T>> HelmholtzDoubleLayerAssembler<T> {
-    /// Create a new assembler
-    pub fn new(wavenumber: T::Real) -> Self {
-        Self {
-            kernel: Helmholtz3dKernel::<T>::new(wavenumber),
-        }
-    }
-}
-impl<T: RlstScalar<Complex = T>> BatchedAssembler for HelmholtzDoubleLayerAssembler<T> {
-    const DERIV_SIZE: usize = 4;
-    type RealT = T::Real;
-    type T = T;
-    unsafe fn singular_kernel_value(
-        &self,
-        k: &RlstArray<Self::T, 2>,
-        _test_normals: &RlstArray<Self::RealT, 2>,
-        trial_normals: &RlstArray<Self::RealT, 2>,
-        index: usize,
-    ) -> Self::T {
-        *k.get_unchecked([1, index])
-            * num::cast::<Self::RealT, Self::T>(*trial_normals.get_unchecked([index, 0])).unwrap()
-            + *k.get_unchecked([2, index])
-                * num::cast::<Self::RealT, Self::T>(*trial_normals.get_unchecked([index, 1]))
-                    .unwrap()
-            + *k.get_unchecked([3, index])
-                * num::cast::<Self::RealT, Self::T>(*trial_normals.get_unchecked([index, 2]))
-                    .unwrap()
-    }
-    unsafe fn nonsingular_kernel_value(
-        &self,
-        k: &RlstArray<Self::T, 3>,
-        _test_normals: &RlstArray<Self::RealT, 2>,
-        trial_normals: &RlstArray<Self::RealT, 2>,
-        test_index: usize,
-        trial_index: usize,
-    ) -> Self::T {
-        *k.get_unchecked([test_index, 1, trial_index])
-            * num::cast::<Self::RealT, Self::T>(*trial_normals.get_unchecked([trial_index, 0]))
-                .unwrap()
-            + *k.get_unchecked([test_index, 2, trial_index])
-                * num::cast::<Self::RealT, Self::T>(*trial_normals.get_unchecked([trial_index, 1]))
-                    .unwrap()
-            + *k.get_unchecked([test_index, 3, trial_index])
-                * num::cast::<Self::RealT, Self::T>(*trial_normals.get_unchecked([trial_index, 2]))
-                    .unwrap()
-    }
-    fn kernel_assemble_diagonal_st(
-        &self,
-        sources: &[Self::RealT],
-        targets: &[Self::RealT],
-        result: &mut [Self::T],
-    ) {
-        self.kernel
-            .assemble_diagonal_st(EvalType::ValueDeriv, sources, targets, result);
-    }
-    fn kernel_assemble_st(
-        &self,
-        sources: &[Self::RealT],
-        targets: &[Self::RealT],
-        result: &mut [Self::T],
-    ) {
-        self.kernel
-            .assemble_st(EvalType::ValueDeriv, sources, targets, result);
-    }
-}
-
-/// Assembler for a Laplace adjoint double layer operator
-pub struct LaplaceAdjointDoubleLayerAssembler<T: RlstScalar> {
-    kernel: Laplace3dKernel<T>,
-}
-impl<T: RlstScalar> Default for LaplaceAdjointDoubleLayerAssembler<T> {
-    fn default() -> Self {
-        Self {
-            kernel: Laplace3dKernel::<T>::new(),
-        }
-    }
-}
-impl<T: RlstScalar> BatchedAssembler for LaplaceAdjointDoubleLayerAssembler<T> {
-    const DERIV_SIZE: usize = 4;
-    type RealT = T::Real;
-    type T = T;
-    unsafe fn singular_kernel_value(
-        &self,
-        k: &RlstArray<T, 2>,
-        test_normals: &RlstArray<<T as RlstScalar>::Real, 2>,
-        _trial_normals: &RlstArray<<T as RlstScalar>::Real, 2>,
-        index: usize,
-    ) -> T {
-        -*k.get_unchecked([1, index])
-            * num::cast::<T::Real, T>(*test_normals.get_unchecked([index, 0])).unwrap()
-            - *k.get_unchecked([2, index])
-                * num::cast::<T::Real, T>(*test_normals.get_unchecked([index, 1])).unwrap()
-            - *k.get_unchecked([3, index])
-                * num::cast::<T::Real, T>(*test_normals.get_unchecked([index, 2])).unwrap()
-    }
-    unsafe fn nonsingular_kernel_value(
-        &self,
-        k: &RlstArray<T, 3>,
-        test_normals: &RlstArray<<T as RlstScalar>::Real, 2>,
-        _trial_normals: &RlstArray<<T as RlstScalar>::Real, 2>,
-        test_index: usize,
-        trial_index: usize,
-    ) -> T {
-        -*k.get_unchecked([test_index, 1, trial_index])
-            * num::cast::<T::Real, T>(*test_normals.get_unchecked([test_index, 0])).unwrap()
-            - *k.get_unchecked([test_index, 2, trial_index])
-                * num::cast::<T::Real, T>(*test_normals.get_unchecked([test_index, 1])).unwrap()
-            - *k.get_unchecked([test_index, 3, trial_index])
-                * num::cast::<T::Real, T>(*test_normals.get_unchecked([test_index, 2])).unwrap()
-    }
-    fn kernel_assemble_diagonal_st(
-        &self,
-        sources: &[T::Real],
-        targets: &[T::Real],
-        result: &mut [T],
-    ) {
-        self.kernel
-            .assemble_diagonal_st(EvalType::ValueDeriv, sources, targets, result);
-    }
-    fn kernel_assemble_st(&self, sources: &[T::Real], targets: &[T::Real], result: &mut [T]) {
-        self.kernel
-            .assemble_st(EvalType::ValueDeriv, sources, targets, result);
     }
 }
